refactor: extract errors to file
