+++ 1 The Physical/Guess Point Level 1 (Stateless)/2 - Arrange-Act-Assert Backwards/2.1 - Password Validator/src/errors.ts
@@ -1,0 +1,6 @@
+export enum Errors {
+    NO_DIGITS = 'NO_DIGITS',
+    TOO_LONG = 'TOO_LONG',
+    TOO_SHORT = 'TOO_SHORT',
+    NO_UPPER = 'NO_UPPER'
+}
