+++ 1 The Physical/Guess Point Level 1 (Stateless)/2 - Arrange-Act-Assert Backwards/2.1 - Password Validator/src/index.spec.ts
@@ -1,4 +1,5 @@
-import { Errors, PasswordValidator } from "./password-validator";
+import { PasswordValidator } from "./password-validator";
+import { Errors } from "./errors";
 
 describe('password validator', () => {
